'value' column in weights changed from integer to float

+++ db/schema.rb
@@ -53,7 +53,7 @@ ActiveRecord::Schema.define(version: 2020_08_17_221548) do
   end
 
   create_table "weights", force: :cascade do |t|
-    t.integer "value", null: false
+    t.float "value", null: false
     t.integer "unity", null: false
     t.date "registration_date"
     t.bigint "user_id", null: false
